refactor(ui): delete useBehavior; migrate alert to the static pattern
The behavior-layer doctrine deleted useBehavior -- every framework file
composes the substrate directly (createBehavior + useMemory). Alert was the
last holdout: an early port still calling useBehavior for {ids, aria}, which
is the only thing that kept the deprecated hook alive.

Alert's score is a pure static (role=alert is constant, no state/effects/ids),
so it takes the card/container shape: config in, classes + aria out, no hook.
keyInputOf (the React KeyboardEvent -> KeyInput translator dialog and
navigation-menu use) moves to its own hooks/key-input.ts; use-behavior.ts is
deleted.
